patch for codegen for Android webview
